--- app/src/main/java/com/msg/gcms/presentation/adapter/AfterSchoolListAdapter.kt
@@ -1,32 +1,0 @@
-package com.msg.gcms.presentation.adapter
-
-import android.view.LayoutInflater
-import android.view.ViewGroup
-import androidx.recyclerview.widget.RecyclerView
-import com.msg.gcms.data.remote.dto.user.response.AftersData
-import com.msg.gcms.databinding.ListAfterSchoolBinding
-
-class AfterSchoolListAdapter(private val itemList: List<AftersData>): RecyclerView.Adapter<AfterSchoolListAdapter.ViewHolder>() {
-    class ViewHolder(val binding: ListAfterSchoolBinding): RecyclerView.ViewHolder(binding.root) {
-        fun bind(after: AftersData) {
-            binding.afterSchoolItem = after
-            binding.gradeTxt.text = "${after.grade}학년"
-        }
-    }
-
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
-        return ViewHolder(
-            ListAfterSchoolBinding.inflate(
-                LayoutInflater.from(parent.context),
-                parent,
-                false
-            )
-        )
-    }
-
-    override fun onBindViewHolder(holder: ViewHolder, position: Int) {
-        holder.bind(itemList[position])
-    }
-
-    override fun getItemCount(): Int = itemList.size
-}
