+++ app/src/main/java/com/msg/gcms/presentation/adapter/add_member/AddMemberAdapter.kt
@@ -1,4 +1,4 @@
-package com.msg.gcms.presentation.adapter
+package com.msg.gcms.presentation.adapter.add_member
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
