+++ app/src/main/java/com/msg/gcms/presentation/adapter/detail_photo/DetailPhotoAdapter.kt
@@ -1,4 +1,4 @@
-package com.msg.gcms.presentation.adapter
+package com.msg.gcms.presentation.adapter.detail_photo
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
