+++ app/src/main/java/com/msg/gcms/presentation/adapter/editorial_club/EditorialClubAdapter.kt
@@ -1,24 +1,23 @@
-package com.msg.gcms.presentation.adapter
+package com.msg.gcms.presentation.adapter.editorial_club
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import coil.load
 import coil.transform.RoundedCornersTransformation
-import com.msg.gcms.databinding.ListClubSummaryBinding
-import com.msg.gcms.domain.data.club.get_club_list.GetClubListData
+import com.msg.gcms.databinding.ListClubEditorialBinding
+import com.msg.gcms.domain.data.user.get_my_profile.ProfileClubData
 
-class ClubListAdapter(private val itemList: List<GetClubListData>?) :
-    RecyclerView.Adapter<ClubListAdapter.ViewHolder>() {
-
-    class ViewHolder(val binding: ListClubSummaryBinding, listener: OnItemClickListener) : RecyclerView.ViewHolder(binding.root) {
-        fun bind(data: GetClubListData?) {
-            binding.clubSummary = data
-            binding.itemClubImg.load(data?.bannerUrl) {
+class EditorialClubAdapter(private val clubList: ArrayList<ProfileClubData>) :
+    RecyclerView.Adapter<EditorialClubAdapter.ViewHolder>() {
+    class ViewHolder(val binding: ListClubEditorialBinding, listener: OnItemClickListener) :
+        RecyclerView.ViewHolder(binding.root) {
+        fun bind(club: ProfileClubData) {
+            binding.editorialItem = club
+            binding.itemClubImg.load(club.bannerImg) {
                 transformations(RoundedCornersTransformation(9f, 9f, 0f, 0f))
             }
         }
-
         init {
             binding.clubSummaryLayout.setOnClickListener {
                 listener.onClick(adapterPosition)
@@ -28,21 +27,19 @@ class ClubListAdapter(private val itemList: List<GetClubListData>?) :
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
         val binding =
-            ListClubSummaryBinding.inflate(LayoutInflater.from(parent.context), parent, false)
+            ListClubEditorialBinding.inflate(LayoutInflater.from(parent.context), parent, false)
         val viewHolder = ViewHolder(binding, itemClickListener)
         return viewHolder
     }
 
     override fun onBindViewHolder(holder: ViewHolder, position: Int) {
-        holder.bind(itemList?.get(position))
+        holder.bind(clubList[position])
         holder.itemView.setOnClickListener {
             itemClickListener.onClick(position)
         }
     }
 
-    override fun getItemCount(): Int {
-        return itemList?.size ?: 0
-    }
+    override fun getItemCount(): Int = clubList.size
 
     interface OnItemClickListener {
         fun onClick(position: Int)
