+++ app/src/main/java/com/msg/gcms/presentation/view/clubmaker/club_detail/MakeClubDetailFragment.kt
@@ -25,8 +25,8 @@ import coil.transform.RoundedCornersTransformation
 import com.msg.gcms.R
 import com.msg.gcms.data.local.entity.ActivityPhotoType
 import com.msg.gcms.databinding.FragmentMakeClubDetailBinding
-import com.msg.gcms.presentation.adapter.ActivityPhotosAdapter
-import com.msg.gcms.presentation.adapter.ClubMemberAdapter
+import com.msg.gcms.presentation.adapter.activity_photo.ActivityPhotosAdapter
+import com.msg.gcms.presentation.adapter.club_member.ClubMemberAdapter
 import com.msg.gcms.presentation.base.BaseFragment
 import com.msg.gcms.presentation.base.BaseModal
 import com.msg.gcms.presentation.utils.ItemDecorator
